Move server code to a subpackage
